Move test factories.
