+++ tests/unit/__init__.py
@@ -2,4 +2,3 @@
 
 from . import helpers
 from . import core
-from . import factories
